+++ crates/header-translator/src/lib.rs
@@ -1,4 +1,5 @@
 #![cfg(feature = "run")]
+#![recursion_limit = "256"]
 use std::collections::BTreeMap;
 use std::fmt;
 use std::path::Path;