+++ crates/icrate/src/CoreAnimation/fixes/mod.rs
@@ -1,0 +1,3 @@
+use std::os::raw::c_double;
+
+pub type CFTimeInterval = c_double;